+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix address lookup not working after changing a source on adding a new location with coordinates input manually (from `geo:` intent)
 
 **Sources**
+- Atmo France - Use new geocoding services to fix refresh issue for some locations
 - China - Fix nowcasting refresh silently failing
 - FMI - Added as a new forecast, current, air quality, alert, normals source for Finland (@chunshek). Users on devices with hardened memory allocator (`hardened_malloc`) enabled, such as GrapheneOS may want to disable the feature for the app if they find the refresh to be too slow.
 - Infoplaza - Added as a new worldwide forecast, current, minutely and normals source (@SKBotNL)